Refactor pipeline base.py & pipelines.py.
