+++ tsdat/pipeline/base/__init__.py
@@ -1,0 +1,5 @@
+from .pipeline import Pipeline
+
+__all__ = [
+    "Pipeline",
+]
