+++ tsdat/pipeline/pipelines/__init__.py
@@ -1,0 +1,8 @@
+from .add_inputs_attr import add_inputs_attr
+from .ingest_pipeline import IngestPipeline
+from .transformation_pipeline import TransformationPipeline
+
+__all__ = [
+    "IngestPipeline",
+    "TransformationPipeline",
+]
